add possibility to interact with storage service, via rabbitmq add possibility to interact with proxy via rabbitmq

+++ src/main/java/kz/company/r_proxy_service/controllers/StudentController.java
@@ -1,6 +1,7 @@
 package kz.company.r_proxy_service.controllers;
 
 import kz.company.r_proxy_service.enums.ResultCode;
+import kz.company.r_proxy_service.exceptions.MethodNotFoundException;
 import kz.company.r_proxy_service.models.dto.ResultMessage;
 import kz.company.r_proxy_service.models.dto.Student;
 import kz.company.r_proxy_service.services.StorageService;
@@ -14,6 +15,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 @RestController
@@ -22,37 +24,52 @@ import java.util.Optional;
 @RequestMapping("/students")
 public class StudentController {
 
-    private final StorageService storageService;
+    private final Map<String, StorageService> storageServiceMap;
 
-    @GetMapping("/{recordBookNumber}")
-    public ResponseEntity<ResultMessage<Student>> getStudent(@PathVariable String recordBookNumber) {
+    @GetMapping("/{method}/{recordBookNumber}")
+    public ResponseEntity<ResultMessage<Student>> getStudent(
+            @PathVariable("method") String method,
+            @PathVariable("recordBookNumber") String recordBookNumber) {
         log.info("StudentController.getStudent request: [{}]", recordBookNumber);
-        Optional<ResultMessage<Student>> resultMessage = storageService.getByRecordBookNumber(recordBookNumber);
-        log.info("StudentController.getStudent response: [{}]", resultMessage);
-        return resultMessage
-                .filter(ResultMessage::isSuccess)
-                .map(ResponseEntity::ok)
-                .or(() -> resultMessage
-                        .filter(rm -> ResultCode.BAD_REQUEST.name().equals(rm.getError().getName()))
-                        .map(rm -> ResponseEntity.badRequest().body(rm))
-                )
-                .orElseGet(() -> resultMessage
-                        .map(rm -> ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(rm))
-                        .orElseGet(() -> ResponseEntity.notFound().build())
-                );
+        if (storageServiceMap.containsKey(method)) {
+            Optional<ResultMessage<Student>> resultMessage = storageServiceMap.get(method)
+                    .getByRecordBookNumber(recordBookNumber);
+            log.info("StudentController.getStudent response: [{}]", resultMessage);
+            return resultMessage
+                    .filter(ResultMessage::isSuccess)
+                    .map(ResponseEntity::ok)
+                    .or(() -> resultMessage
+                            .filter(rm -> ResultCode.BAD_REQUEST.name().equals(rm.getError().getName()))
+                            .map(rm -> ResponseEntity.badRequest().body(rm))
+                    )
+                    .orElseGet(() -> resultMessage
+                            .map(rm -> ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(rm))
+                            .orElseGet(() -> ResponseEntity.notFound().build())
+                    );
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body(ResultMessage.error(
+                            new MethodNotFoundException(method).getMessage()));
+        }
     }
 
-    @GetMapping()
-    public ResponseEntity<ResultMessage<List<Student>>> getStudents() {
+    @GetMapping("/{method}")
+    public ResponseEntity<ResultMessage<List<Student>>> getStudents(@PathVariable("method") String method) {
         log.info("StudentController.getStudents request");
-        Optional<ResultMessage<List<Student>>> resultMessage = storageService.getAll();
-        log.info("StudentController.getStudents response: [{}]", resultMessage);
-        return resultMessage
-                .filter(ResultMessage::isSuccess)
-                .map(ResponseEntity::ok)
-                .orElseGet(() -> resultMessage
-                        .map(rm -> ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(rm))
-                        .orElseGet(() -> ResponseEntity.notFound().build())
-                );
+        if (storageServiceMap.containsKey(method)) {
+            Optional<ResultMessage<List<Student>>> resultMessage = storageServiceMap.get(method).getAll();
+            log.info("StudentController.getStudents response: [{}]", resultMessage);
+            return resultMessage
+                    .filter(ResultMessage::isSuccess)
+                    .map(ResponseEntity::ok)
+                    .orElseGet(() -> resultMessage
+                            .map(rm -> ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(rm))
+                            .orElseGet(() -> ResponseEntity.notFound().build())
+                    );
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body(ResultMessage.error(
+                            new MethodNotFoundException(method).getMessage()));
+        }
     }
 }
